+++ lib/ChargeBee/Models/Invoice.php
@@ -36,6 +36,16 @@ class ChargeBee_Invoice extends ChargeBee_Model
     return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices","import_invoice"), $params, $env, $headers);
   }
 
+  public static function applyPayments($id, $params = array(), $env = null, $headers = array())
+  {
+    return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices",$id,"apply_payments"), $params, $env, $headers);
+  }
+
+  public static function applyCredits($id, $params = array(), $env = null, $headers = array())
+  {
+    return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices",$id,"apply_credits"), $params, $env, $headers);
+  }
+
   public static function all($params = array(), $env = null, $headers = array())
   {
     return ChargeBee_Request::sendListRequest(ChargeBee_Request::GET, ChargeBee_Util::encodeURIPath("invoices"), $params, $env, $headers);
@@ -94,6 +104,16 @@ class ChargeBee_Invoice extends ChargeBee_Model
   public static function recordRefund($id, $params, $env = null, $headers = array())
   {
     return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices",$id,"record_refund"), $params, $env, $headers);
+  }
+
+  public static function removePayment($id, $params, $env = null, $headers = array())
+  {
+    return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices",$id,"remove_payment"), $params, $env, $headers);
+  }
+
+  public static function removeCreditNote($id, $params, $env = null, $headers = array())
+  {
+    return ChargeBee_Request::send(ChargeBee_Request::POST, ChargeBee_Util::encodeURIPath("invoices",$id,"remove_credit_note"), $params, $env, $headers);
   }
 
   public static function voidInvoice($id, $params = array(), $env = null, $headers = array())